test(ui): narrow the SDL key modifier explicitly
SDL_Keysym.mod is a Uint16 while SDL_Keymod is an enum, and clang's
-Wimplicit-int-conversion flagged the assignment in sdl_push_key().

+++ tests/suites/ui_sdl.c
@@ -547,7 +547,7 @@ static void sdl_push_key(SDL_Scancode scancode, SDL_Keycode symbol, SDL_Keymod m
     event.key.type = SDL_KEYDOWN;
     event.key.keysym.scancode = scancode;
     event.key.keysym.sym = symbol;
-    event.key.keysym.mod = modifiers;
+    event.key.keysym.mod = (Uint16)modifiers;
     event.key.repeat = repeat;
     SDL_PushEvent(&event);
 }
